Add React hook
